Refactor D3D12MA to use the new TerraFX split
