+++ TerraFX.Interop.D3D12MemoryAllocator.sln
@@ -82,6 +82,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "generation", "generation", 
 		generation\settings.rsp = generation\settings.rsp
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TerraFX.Interop.Windows", "..\terrafx.interop.windows\sources\Interop\Windows\TerraFX.Interop.Windows.csproj", "{45BA1430-F6F4-4A23-BBB1-EC9A5081DEFA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -96,6 +98,10 @@ Global
 		{FE1214EA-21BD-4087-AD01-D42DBA8572FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FE1214EA-21BD-4087-AD01-D42DBA8572FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FE1214EA-21BD-4087-AD01-D42DBA8572FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45BA1430-F6F4-4A23-BBB1-EC9A5081DEFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45BA1430-F6F4-4A23-BBB1-EC9A5081DEFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45BA1430-F6F4-4A23-BBB1-EC9A5081DEFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45BA1430-F6F4-4A23-BBB1-EC9A5081DEFA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
