+++ sources/Interop/D3D12MemoryAllocator/inc/D3D12MemAlloc/D3D12MA_ALLOCATOR_DESC.cs
@@ -3,6 +3,10 @@
 // Ported from D3D12MemAlloc.h in D3D12MemoryAllocator commit 5457bcdaee73ee1f3fe6027bbabf959119f88b3d
 // Original source is Copyright © Advanced Micro Devices, Inc. All rights reserved. Licensed under the MIT License (MIT).
 
+using TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows.D3D12;
+using TerraFX.Interop.Windows.DXGI;
+
 namespace TerraFX.Interop.D3D12MA
 {
     /// <summary>Parameters of created <see cref="D3D12MA_Allocator"/> object. To be used with <see cref="D3D12MemAlloc.D3D12MA_CreateAllocator"/>.</summary>
