+++ sources/Interop/D3D12MemoryAllocator/inc/D3D12MemAlloc/D3D12MA_Allocation.cs
@@ -6,9 +6,11 @@
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
-using static TerraFX.Interop.D3D12_RESOURCE_DIMENSION;
-using static TerraFX.Interop.D3D12_RESOURCE_FLAGS;
-using static TerraFX.Interop.D3D12_TEXTURE_LAYOUT;
+using TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows.D3D12;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_DIMENSION;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_TEXTURE_LAYOUT;
 using static TerraFX.Interop.D3D12MA.D3D12MemAlloc;
 using static TerraFX.Interop.D3D12MA.D3D12MA_Allocation.Type;
 using static TerraFX.Interop.D3D12MA.D3D12MA_Allocation._Anonymous_e__Union;
