+++ sources/Interop/D3D12MemoryAllocator/inc/D3D12MemAlloc/D3D12MA_POOL_DESC.cs
@@ -3,8 +3,9 @@
 // Ported from D3D12MemAlloc.h in D3D12MemoryAllocator commit 5457bcdaee73ee1f3fe6027bbabf959119f88b3d
 // Original source is Copyright © Advanced Micro Devices, Inc. All rights reserved. Licensed under the MIT License (MIT).
 
-using static TerraFX.Interop.Windows;
-using static TerraFX.Interop.D3D12_HEAP_FLAGS;
+using TerraFX.Interop.Windows.D3D12;
+using static TerraFX.Interop.Windows.Windows;
+using static TerraFX.Interop.Windows.D3D12.D3D12_HEAP_FLAGS;
 
 namespace TerraFX.Interop.D3D12MA
 {
