+++ sources/Interop/D3D12MemoryAllocator/inc/D3D12MemAlloc/D3D12MA_VirtualBlock.cs
@@ -6,7 +6,7 @@
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
-using static TerraFX.Interop.Windows;
+using static TerraFX.Interop.Windows.Windows;
 using static TerraFX.Interop.D3D12MA.D3D12MemAlloc;
 
 namespace TerraFX.Interop.D3D12MA
