+++ sources/Interop/D3D12MemoryAllocator/src/D3D12MemAlloc/D3D12MA_BlockMetadata_Generic.cs
@@ -5,7 +5,7 @@
 
 using System;
 using System.Runtime.CompilerServices;
-using static TerraFX.Interop.Windows;
+using static TerraFX.Interop.Windows.Windows;
 using static TerraFX.Interop.D3D12MA.D3D12MemAlloc;
 using static TerraFX.Interop.D3D12MA.D3D12MA_SuballocationType;
