+++ sources/Interop/D3D12MemoryAllocator/src/D3D12MemAlloc/D3D12MA_CommittedAllocationList.cs
@@ -5,7 +5,8 @@
 
 using System;
 using System.Runtime.CompilerServices;
-using static TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows.D3D12;
+using static TerraFX.Interop.Windows.Windows;
 using static TerraFX.Interop.D3D12MA.D3D12MemAlloc;
 
 namespace TerraFX.Interop.D3D12MA
