+++ sources/Interop/D3D12MemoryAllocator/src/D3D12MemAlloc/D3D12MA_MUTEX.cs
@@ -4,7 +4,8 @@
 // Original source is Copyright © Advanced Micro Devices, Inc. All rights reserved. Licensed under the MIT License (MIT).
 
 using System;
-using static TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows;
+using static TerraFX.Interop.Windows.Windows;
 
 namespace TerraFX.Interop.D3D12MA
 {
