+++ sources/Interop/D3D12MemoryAllocator/src/D3D12MemAlloc/D3D12MA_PoolAllocator`1.cs
@@ -6,6 +6,7 @@
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using static TerraFX.Interop.Windows.Windows;
 using static TerraFX.Interop.D3D12MA.D3D12MemAlloc;
 
 namespace TerraFX.Interop.D3D12MA
