+++ tests/Interop/D3D12MemoryAllocator/inc/Tests/TestContext.cs
@@ -3,6 +3,8 @@
 // Ported from Tests.h in D3D12MemoryAllocator commit 5457bcdaee73ee1f3fe6027bbabf959119f88b3d
 // Original source is Copyright © Advanced Micro Devices, Inc. All rights reserved. Licensed under the MIT License (MIT).
 
+using TerraFX.Interop.Windows.D3D12;
+
 namespace TerraFX.Interop.D3D12MA.UnitTests
 {
     internal unsafe struct TestContext
