+++ tests/Interop/D3D12MemoryAllocator/src/D3D12Sample/D3D12MemAllocTests.cs
@@ -7,47 +7,51 @@ using System;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
-using static TerraFX.Interop.D3D_FEATURE_LEVEL;
-using static TerraFX.Interop.D3D_PRIMITIVE_TOPOLOGY;
-using static TerraFX.Interop.D3D_ROOT_SIGNATURE_VERSION;
-using static TerraFX.Interop.D3D12_BLEND;
-using static TerraFX.Interop.D3D12_BLEND_OP;
-using static TerraFX.Interop.D3D12_CLEAR_FLAGS;
-using static TerraFX.Interop.D3D12_COLOR_WRITE_ENABLE;
-using static TerraFX.Interop.D3D12_COMMAND_LIST_TYPE;
-using static TerraFX.Interop.D3D12_COMPARISON_FUNC;
-using static TerraFX.Interop.D3D12_CONSERVATIVE_RASTERIZATION_MODE;
-using static TerraFX.Interop.D3D12_CULL_MODE;
-using static TerraFX.Interop.D3D12_DEPTH_WRITE_MASK;
-using static TerraFX.Interop.D3D12_DESCRIPTOR_HEAP_FLAGS;
-using static TerraFX.Interop.D3D12_DESCRIPTOR_HEAP_TYPE;
-using static TerraFX.Interop.D3D12_DESCRIPTOR_RANGE_TYPE;
-using static TerraFX.Interop.D3D12_DSV_DIMENSION;
-using static TerraFX.Interop.D3D12_DSV_FLAGS;
-using static TerraFX.Interop.D3D12_FENCE_FLAGS;
-using static TerraFX.Interop.D3D12_FILL_MODE;
-using static TerraFX.Interop.D3D12_FILTER;
-using static TerraFX.Interop.D3D12_HEAP_TYPE;
-using static TerraFX.Interop.D3D12_INPUT_CLASSIFICATION;
-using static TerraFX.Interop.D3D12_LOGIC_OP;
-using static TerraFX.Interop.D3D12_PRIMITIVE_TOPOLOGY_TYPE;
-using static TerraFX.Interop.D3D12_RESOURCE_BARRIER_TYPE;
-using static TerraFX.Interop.D3D12_RESOURCE_DIMENSION;
-using static TerraFX.Interop.D3D12_RESOURCE_FLAGS;
-using static TerraFX.Interop.D3D12_RESOURCE_HEAP_TIER;
-using static TerraFX.Interop.D3D12_RESOURCE_STATES;
-using static TerraFX.Interop.D3D12_ROOT_PARAMETER_TYPE;
-using static TerraFX.Interop.D3D12_ROOT_SIGNATURE_FLAGS;
-using static TerraFX.Interop.D3D12_SHADER_VISIBILITY;
-using static TerraFX.Interop.D3D12_SRV_DIMENSION;
-using static TerraFX.Interop.D3D12_STATIC_BORDER_COLOR;
-using static TerraFX.Interop.D3D12_STENCIL_OP;
-using static TerraFX.Interop.D3D12_TEXTURE_ADDRESS_MODE;
-using static TerraFX.Interop.D3D12_TEXTURE_LAYOUT;
-using static TerraFX.Interop.DXGI_ADAPTER_FLAG;
-using static TerraFX.Interop.DXGI_FORMAT;
-using static TerraFX.Interop.DXGI_SWAP_EFFECT;
-using static TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows.D3D12;
+using TerraFX.Interop.Windows.D3DCommon;
+using TerraFX.Interop.Windows.DXGI;
+using static TerraFX.Interop.Windows.D3DCommon.D3D_FEATURE_LEVEL;
+using static TerraFX.Interop.Windows.D3DCommon.D3D_PRIMITIVE_TOPOLOGY;
+using static TerraFX.Interop.Windows.D3D12.D3D_ROOT_SIGNATURE_VERSION;
+using static TerraFX.Interop.Windows.D3D12.D3D12_BLEND;
+using static TerraFX.Interop.Windows.D3D12.D3D12_BLEND_OP;
+using static TerraFX.Interop.Windows.D3D12.D3D12_CLEAR_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_COLOR_WRITE_ENABLE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_COMMAND_LIST_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_COMPARISON_FUNC;
+using static TerraFX.Interop.Windows.D3D12.D3D12_CONSERVATIVE_RASTERIZATION_MODE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_CULL_MODE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_DEPTH_WRITE_MASK;
+using static TerraFX.Interop.Windows.D3D12.D3D12_DESCRIPTOR_HEAP_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_DESCRIPTOR_HEAP_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_DESCRIPTOR_RANGE_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_DSV_DIMENSION;
+using static TerraFX.Interop.Windows.D3D12.D3D12_DSV_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_FENCE_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_FILL_MODE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_FILTER;
+using static TerraFX.Interop.Windows.D3D12.D3D12_HEAP_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_INPUT_CLASSIFICATION;
+using static TerraFX.Interop.Windows.D3D12.D3D12_LOGIC_OP;
+using static TerraFX.Interop.Windows.D3D12.D3D12_PRIMITIVE_TOPOLOGY_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_BARRIER_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_DIMENSION;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_HEAP_TIER;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_STATES;
+using static TerraFX.Interop.Windows.D3D12.D3D12_ROOT_PARAMETER_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_ROOT_SIGNATURE_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_SHADER_VISIBILITY;
+using static TerraFX.Interop.Windows.D3D12.D3D12_SRV_DIMENSION;
+using static TerraFX.Interop.Windows.D3D12.D3D12_STATIC_BORDER_COLOR;
+using static TerraFX.Interop.Windows.D3D12.D3D12_STENCIL_OP;
+using static TerraFX.Interop.Windows.D3D12.D3D12_TEXTURE_ADDRESS_MODE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_TEXTURE_LAYOUT;
+using static TerraFX.Interop.Windows.DXGI.DXGI_ADAPTER_FLAG;
+using static TerraFX.Interop.Windows.DXGI.DXGI_FORMAT;
+using static TerraFX.Interop.Windows.DXGI.DXGI_SWAP_EFFECT;
+using static TerraFX.Interop.Windows.Windows;
 using static TerraFX.Interop.D3D12MA.D3D12MA_ALLOCATOR_FLAGS;
 using static TerraFX.Interop.D3D12MA.D3D12MemAlloc;
