+++ tests/Interop/D3D12MemoryAllocator/src/Tests/D3D12MemAllocTests.cs
@@ -9,19 +9,21 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading;
 using NUnit.Framework;
-using static TerraFX.Interop.D3D12_CPU_PAGE_PROPERTY;
-using static TerraFX.Interop.D3D12_FEATURE;
-using static TerraFX.Interop.D3D12_HEAP_FLAGS;
-using static TerraFX.Interop.D3D12_HEAP_TYPE;
-using static TerraFX.Interop.D3D12_MEMORY_POOL;
-using static TerraFX.Interop.D3D12_PROTECTED_RESOURCE_SESSION_SUPPORT_FLAGS;
-using static TerraFX.Interop.D3D12_RESOURCE_BARRIER_TYPE;
-using static TerraFX.Interop.D3D12_RESOURCE_DIMENSION;
-using static TerraFX.Interop.D3D12_RESOURCE_FLAGS;
-using static TerraFX.Interop.D3D12_RESOURCE_STATES;
-using static TerraFX.Interop.D3D12_TEXTURE_LAYOUT;
-using static TerraFX.Interop.DXGI_FORMAT;
-using static TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows;
+using TerraFX.Interop.Windows.D3D12;
+using static TerraFX.Interop.Windows.D3D12.D3D12_CPU_PAGE_PROPERTY;
+using static TerraFX.Interop.Windows.D3D12.D3D12_FEATURE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_HEAP_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_HEAP_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_MEMORY_POOL;
+using static TerraFX.Interop.Windows.D3D12.D3D12_PROTECTED_RESOURCE_SESSION_SUPPORT_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_BARRIER_TYPE;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_DIMENSION;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_FLAGS;
+using static TerraFX.Interop.Windows.D3D12.D3D12_RESOURCE_STATES;
+using static TerraFX.Interop.Windows.D3D12.D3D12_TEXTURE_LAYOUT;
+using static TerraFX.Interop.Windows.DXGI.DXGI_FORMAT;
+using static TerraFX.Interop.Windows.Windows;
 using static TerraFX.Interop.D3D12MA.D3D12MA_ALLOCATION_FLAGS;
 using static TerraFX.Interop.D3D12MA.D3D12MA_ALLOCATOR_FLAGS;
 using static TerraFX.Interop.D3D12MA.D3D12MemAlloc;
